make suite_config a dataclass
