+++ singer_sdk/testing/suites.py
@@ -42,23 +42,6 @@ from .templates import TapTestTemplate, TargetTestTemplate, TestTemplate
 
 
 @dataclass
-class SuiteConfig:
-    """Test Suite Config, passed to each test.
-
-    Args:
-        max_records_limit: Max records to fetch during tap testing.
-        ignore_no_records: Ignore stream test failures if stream returns no records,
-            for all streams.
-        ignore_no_records_for_streams: Ignore stream test failures if stream returns
-             no records, for named streams.
-    """
-
-    max_records_limit: int | None = 5
-    ignore_no_records: bool = False
-    ignore_no_records_for_streams: list[str] = []
-
-
-@dataclass
 class TestSuite:
     """Test Suite container class."""
